+++ migrations/000003_create_loyalty_points_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE loyalty_points (
     user_id UUID PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
-    balance INT DEFAULT 0,
+    balance FLOAT DEFAULT 0,
     created_at TIMESTAMP DEFAULT NOW(),
     updated_at TIMESTAMP DEFAULT NOW()
 );
